Move on_event tests into test_run.py
